Changes names function to Index
