+++ client/src/components/ContentModal/Grocery/ContentGrocery.jsx
@@ -1,45 +1,85 @@
-import { useMemo } from "react";
-import { useState, useEffect } from "react";
-import Index from "./Index.jsx";
-import { activitiesUser, products } from "../../../utils/data.js";
+import { Tabs, Tab } from "@heroui/react";
+import React, { useState } from "react";
+import ProductsTab from "./Products.jsx";
+import ShoppingListTab from "./Shopping.jsx";
+import AtmTab from "./ATM.jsx";
+import {
+  CreditCardIcon,
+  ListBulletIcon,
+  ShoppingCartIcon,
+} from "@heroicons/react/24/solid";
 
-export default function ContentGrocery({ statusCard }) {
-  const itemsPerPage = 8;
-  const [currentPage, setCurrentPage] = useState(1);
-  
-  const displayedItems = useMemo(() => {
-    return statusCard === "activitiesUser" ? activitiesUser : products;
-  }, [statusCard]);
+const ContentGrocery = React.memo(
+  ({ itemsToDisplay, page, total, onChange }) => {
+    // State Management
+    const [paymentStatus, setPaymentStatus] = useState("Make Payment");
+    const [paymentProcessing, setPaymentProcessing] = useState(false);
+    const [alertVisible, setAlertVisible] = useState(false);
+    const [alertType, setAlertType] = useState("success"); // success or danger
+    const [selectedProducts, setSelectedProducts] = useState([]);
 
-  // Reset the current page when statusCard changes
-  useEffect(() => {
-    setCurrentPage(1);
-  }, [statusCard]);
+    return (
+      <Tabs
+        aria-label="Options"
+        variant="solid"
+        color="primary"
+        size="lg"
+        fullWidth
+      >
+        <Tab
+          key="products"
+          title={
+            <div className="flex items-center space-x-2">
+              <ShoppingCartIcon className="size-6" />
+              <span>Products</span>
+            </div>
+          }
+        >
+          <ProductsTab
+            itemsToDisplay={itemsToDisplay}
+            page={page}
+            total={total}
+            onChange={onChange}
+            selectedProducts={selectedProducts}
+            setSelectedProducts={setSelectedProducts}
+          />
+        </Tab>
 
-  const indexOfLastItem = currentPage * itemsPerPage;
-  const indexOfFirstItem = indexOfLastItem - itemsPerPage;
-  const currentItems = useMemo(() => {
-    return displayedItems.slice(indexOfFirstItem, indexOfLastItem);
-  }, [displayedItems, indexOfFirstItem, indexOfLastItem]);
-  
+        <Tab
+          key="shoppinglist"
+          title={
+            <div className="flex items-center space-x-2">
+              <ListBulletIcon className="size-6" />
+              <span>Shopping List</span>
+            </div>
+          }
+        >
+          <ShoppingListTab selectedProducts={selectedProducts} />
+        </Tab>
 
-  const handlePageChange = (newPage) => {
-    setCurrentPage(newPage);
-  };
+        <Tab
+          key="atm"
+          title={
+            <div className="flex items-center space-x-2">
+              <CreditCardIcon className="size-6" />
+              <span>ATM (Cashier)</span>
+            </div>
+          }
+        >
+          <AtmTab
+            paymentStatus={paymentStatus}
+            setPaymentStatus={setPaymentStatus}
+            paymentProcessing={paymentProcessing}
+            setPaymentProcessing={setPaymentProcessing}
+            alertVisible={alertVisible}
+            setAlertVisible={setAlertVisible}
+            alertType={alertType}
+            setAlertType={setAlertType}
+          />
+        </Tab>
+      </Tabs>
+    );
+  },
+);
 
-  if (!Array.isArray(displayedItems)) {
-    console.error("Error: displayedItems is not an array", displayedItems);
-    return <p>No products available to display.</p>;
-  }
-
-  const totalPages = Math.ceil(displayedItems.length / itemsPerPage);
-
-  return (
-    <Index
-      itemsToDisplay={currentItems}
-      page={currentPage}
-      total={totalPages}
-      onChange={handlePageChange}
-    />
-  );
-}
+export default ContentGrocery;
